[v1.19.0-step-1] CompositingView 1366줄 → 6개 파일 분할 (동작 동일)
